Curriculum Pane Updates (#23)
* Scroll Table of Contents, search results.

Trickier than just slapping on `overflow-y: scroll` because they are
`position: absolute`.

This commit introduces a reusable hook-esque thing, `useHeightLimiter`,
for ensuring that such elements do not exceed the screen height (and
instead get a scrollbar). This might also come in handy beyond the
curriculum pane.

* Trigger language change in Angular components.

* Replace console.log() with esconsole().

Avoids spamming the console with debug messages.

* Spruce up the curriculum loading screen.

Now uses a scaled down version of the EarSketch loading spinner.

* Fix intra-curricular links, refactor processing.

Previously, cross-references were implemented by adding onclick strings
to the HTML, to invoke angular functions. This took place during
curriculum preprocessing in Python.

Now, this happens in the client with the other DOM processing steps, and
the event handlers are added by JS (and thus not visible in the HTML).
This required changing how we process and load the curriculum pages.
Also, this change finally moves the processing out of
fetchContent.fulfilled, as I promised @ttsuchiya a while ago, and it
removes the now-defunct maximization code from curriculumState.

* Start updating content CSS, fix hilitor bug.

* Refactor questions.js, update styling.

Great pains were taken to make everything wrap correctly.
Might play around with this a little more.

* More question styling, fix scripts re-running.

* Comment defunct styles, integrate with Tailwind.

Work on using Tailwind Typography, as recommended by @ttsuchiya, and
using Tailwind's support for dark mode.

Still need to identify which styles we'll keep before throwing them out.

* remove allstyles.css

* add grunt less to local dev builds

* Remove unused curriculum CSS.

Odds are good we'll end up bringing some of this back, but things are
looking pretty good with tailwind typography and the remaining rules.

Co-authored-by: Stephen Garrett <garrett@gatech.edu>