Autograder Updates Spring 2024 (#639)
In compliance with recent changes to React and EarSketch:

- `sourceLine` comparison: we remove the `sourceLine` field (used for
DAW -> Editor highlighting) when comparing reference files to test
files, since this was causing identical outputs to fail based on spacing
differences.
- `useEffect` cleanup: Because useEffect applies twice, we now destroy
CodeMirror contents to clean up the `CodeEmbed` object. We do this with
`container`, a local copy of `editorContainer.current`.
- `results` mutation: we now copy the `results` object locally, so that
the `TestResults` updates correctly instead of "freezing".