Handle all dates as iso 8601 strings (#696)
Removes the special handling of non-standard date formatting.

Fixes notifications, script history "NaN years ago" with the updated
backend api.

Fixes todo for using a standard date format.
> TODO: Dates should be stored in a standard format in the database so
as to make this unnecessary.

This is needed for our database update to MySQL 8. Relates to
GTCMT/earsketch-api#84, which configures datetime fields to serialize as
standard ISO 8601 date strings (yay!). The special handling of the old
format breaks on the standard format.