Remove collaboration mode (#727)
The option for collaborative editing (share -> let others edit) is now disabled, and existing collaborative scripts will be treated as regular scripts for the owner and shared scripts for the collaborators.

This does not require a corresponding server-side change or database update. Eventually, we will want to convert collaborative scripts to shared scripts. In the meantime, they should appear and behave as shared scripts in the client.

Resolves GTCMT/earsketch#3235.

---------

Co-authored-by: Luke Heerman <43624284+heerman@users.noreply.github.com>
Co-authored-by: Stephen Garrett <manodrum@gmail.com>