Restore autograder entrypoint in Vite config (#750)
Autograder can once again be accessed via `/autograder`.

Though, I'm not sure we even need a separate entrypoint for this;
`index_autograders.html` (now `autograder/index.html`) is almost
identical to `index.html`. The only real difference is that it's served
from `/autograder`, which the web client checks from in `index.tsx`. But
that same could be accomplished with a copy or maybe even a symbolic
link.

(This is assuming we need to preserve the old URL; if we don't, we could
also just switch to a query parameter.)